// src/interfaces/planet.dto.tsx
export interface PlanetDTO {
  readonly climate: string;
  readonly created: string;
  readonly diameter: string;
  readonly edited: string;
  readonly films: string[];
  readonly gravity: string;
  readonly name: string;
  readonly orbital_period: string;
  readonly population: string;
  readonly residents: string[];
  readonly rotation_period: string;
  readonly surface_water: string;
  readonly terrain: string;
  readonly url: string;
}